+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - **Breaking**: generic: 最小サポートを2.1.2に引き上げ
 - **Breaking**: generic: `ReadSection::get_object_section_frame`の戻り値を`EditSectionResult<usize>`に
+- **Breaking**: filter: トレイトに`Userdata`型を追加
+- **Breaking**: filter: FilterVideoProc・FilterAudioProcに`Userdata`ジェネリクスを追加
 - input: 高速なAPIを追加
 - generic: ReadSectionProviderをpubにする
 - eframe: フォント読み込みで大文字小文字違いを許容するようにする